requirements.txt
httpx
python-dotenv
tqdm
colorama
psutil
python-telegram-bot
Jinja2
tenacity
anyio==4.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psutil==7.0.0
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.1.1
requests==2.32.4
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
